Improved the help given when the actor throws an exception due to not being in the world.
One aspect is that it now indicates whether it has never been in a world, or whether it was in one, but has been removed.  But the more useful addition is that if the actor has been removed, you get a stack trace in the exception's "caused by", showing when it was removed.

Fixes GREENFOOT-371