Added a new JavaParserCallbacks superclass for JavaParser, and moved all the overridable callback methods up into it.
Previously, these methods were scattered throughout JavaParser and there wasn't a clear distinction between the overridable methods and the ones that should not be overridden.  Now all the overridable ones are in the parent class, and all the JavaParser methods (bar one or two special cases) are now final, to indicate they shouldn't be overridden.