Improve handling of non-English compiler error messages.
Previously, we would fetch compiler error messages in the default locale, i.e. the machine's locale.  This meant that when the JDK had support, the compiler error messages would be non-English.  However, our support for quick fixes relies on English messages, so it wouldn't work if the compiler errors were localised by the JDK.

With this change, we actually fetch in English and localised in the user's specified BlueJ interface language (which may or may not be English, or the machine's locale) -- this may fall back to English anyway if the JDK has no support for that language.  We generally display the hopefully-localised version of the error messages, but we use the English version for the quick fixes.  In this case we reword the error messages which means they will be localised if we have localised text for them, otherwise they will be English.

This means some locales, which support JDK localisation but don't have BlueJ localisation for the reworded errors, may see a mix of localised messages without quick fixes, and English messages with quick fixes.  But my sense is that is preferable to not having the quick fixes.  This currently applies to German, for example, and Japanese.  But it could be fixed by adding the localisations to BlueJ in the future.